Interní symboly (EN)

Internal Documentation

The following are the private symbols from the module MarkdownChains.

MarkovChains.build — Method.

build(suptokens::Vector{Vector{Token{T}}}; order=2, weight=stdweight)

Trains a Markov chain on an array of arrays of tokens (suptokens). Optionally an order of the chain can be supplied, that is the number of tokens in one state. A weight function of general type func(s::State{T}, t::Token{T})::Int can be supplied to be used to bias the weights based on the state or token.

source

MarkovChains.combine — Method.

function combine(chain::Model, others::Model...)

Return a Model which is a combination of all of the models provided. All of the arguments should have the same order. The nodes of all the Models are merged using the function merge.

source

MarkovChains.state_with_beginning — Method.

state_with_beginning(model::Model{T}, tokens::Vector{Token{T}}; strict=false) where T

Attempts to return a random valid state of model that begins with tokens. If strict is false and the model doesn't have any state that begins with tokens, the function shortens the tokens (cuts the last token) to lower the requirements and tries to find some valid state again.

source

MarkovChains.walk — Method.

walk(model::Model)

Return an array of tokens obtained by a random walk through the Markov chain. The walk starts at state [:begin, :begin...] (the length depends on the order of the supplied model) and ends once a special token :end is reached.

See also: walk2.

source

MarkovChains.walk — Method.

walk(model::Model{T}[, init_state::State{T}]) where T

Return an array of tokens obtained by a random walk through the Markov chain. The walk starts at state init_state and ends once a special token :end is reached.

See also: walk2.

source

MarkovChains.walk2 — Method.

walk2(model::Model{Any})

Return an array of tokens obtained by a random walk through the Markov chain. When there is only one state following the current one (i.e. there is 100% chance that the state will become the next one), the function shortens the current State as to lower the requirements and obtain more randomness. The State gets shortened until a state with at least two possible successors is found (or until State is only one token long).

The walk starts at state [:begin, :begin...] (the length depends on the order of the supplied model) and ends once a special token :end is reached.

See also: walk.

source

MarkovChains.walk2 — Method.

walk2(model::Model{T}[, init_state::State{T}]) where T

Return an array of tokens obtained by a random walk through the Markov chain. When there is only one state following the current one (i.e. there is 100% chance that the state will become the next one), the function shortens the current State as to lower the requirements and obtain more randomness. The State gets shortened until a state with at least two possible successors is found (or until State is only one token long).

The walk starts at state init_state and ends once a special token :end is reached.

See also: walk.

source

MarkovChains.Model — Type.

The datastructure of the Markov chain. Encodes all the different states and the probabilities of going from one to another as a dictionary. The keys are the states, the values are the respective TokenOccurences dictionaries. Those are dictionaries which say how many times was a token found immediately after the state.

Fields

order is the number of tokens the State

nodes is a dictionary pairing State and its respective

TokenOccurences dictionary.

source

MarkovChains.State — Type.

State{T} = Vector{Token{T}}

A state is a succession of tokens.

source

MarkovChains.TokenOccurences — Type.

TokenOccurences{Token} = Dict{Union{Token, Symbol}, Int}

A dictionary pairing tokens (or special symbols :begin and :end) with the number of their respective occurences.

source

MarkovChains.append_token — Method.

append_token(state::State{T}, token::Token{T}) where T

Drop the first element in state and append the token at the end of the state array.

source

MarkovChains.begseq — Method.

begseq(n)

Return the symbol :begin repeated n times. This array is then used as a starting sequence for all suptokens.

source

MarkovChains.indexof — Method.

indexof(array::AbstractVector{T}, n::T) where T

Given a sorted array, return the index on which n would be inserted in should the insertion preserve the sorting.

source

MarkovChains.next_token — Method.

next_token(model::Model{T}, state::State{T}) where T

Return a token which will come after the current state, at random. The probabilities of individual tokens getting choosed are skewed by their individual values in the TokenOccurences dictionary of the current state, that is obtained from the model.

source

MarkovChains.randkey — Method.

randkey(dict::AbstractDict{Any, Number})

Return a random key from dict. The probabilities of individual keys getting chosen are skewed by their respective values.

source

MarkovChains.states_with_suffix — Method.

states_with_suffix(model::Model{T}, init_suffix::Vector{Tokens{T}}) where T

Return all of the states of model that end with init_suffix. If the number of such states is 1 (or 0), the function shortens the suffix (cuts the first token) in order to lower the requirements, and makes another try.

source

MarkovChains.stdweight — Method.

stdweight(state::State{T}, token::Token{T}) where T

A constant 1. Used as a placeholder function in build to represent unbiased weight function.

source

MarkovChains.walker — Method.

walker(model::Model{T}, init_state::State{T}, init_accum, newstate=append_token) where T

Return an array of tokens obtained by a random walk through the Markov chain. The walk starts at state init_state and ends once a special token :end is reached. A function newstate of general type func(state::State{T}, token::Token{T})::State{T} where T can be supplied to be used to generate a new state given the old state and the following token.

This is a general function which is used by all the walk functions.

See also: walk, walk2.

source

The following are the private symbols from the module Tokenizer.

Tokenizer.cleanup — Method.

cleanup(suptokens::Vector{Vector{String}}; badchars="

-_()[]{}<>–—$='"„“ ")

Remove all characters that are in badchars from all tokens in suptokens.

source

Tokenizer.letters — Function.

letters = cleanup ∘ to_letters ∘ to_sentences

Composite function which splits its input into sentences, then the sentences into letters, and then removes special characters.

source

Tokenizer.lines — Function.

lines = cleanup ∘ to_letters ∘ to_sentences

Composite function which splits its input into lines, then the line into letters, and then removes special characters.

source

Tokenizer.to_letters — Method.

to_letters(tokens::Vector{String})

Split all of the tokens in tokens into individual characters.

source

Tokenizer.to_lines — Method.

to_lines(text)

Return an array of lines in text.

source

Tokenizer.to_sentences — Method.

to_sentences(text)

Return an array of sentences in text. The text is split along dots; the dots remain in the strings, only the spaces after the dots are stripped.

The function tries to be as smart as possible. For example, the string "Channel No. 5 is a perfume." will be treated as one sentence, although it has two dots.

source

Tokenizer.tokenize — Function.

tokenize(inp[, func=letters])

Split the text into SupTokens (list of lists of tokens). An optional function of general type func(inp:T1)::Vector{Vector{T2}} can be provided to be used for the tokenization.

For possible combinators which can be composed to obtain func, see: to_lines, to_sentences, to_letters, to_words, cleanup.

source

Tokenizer.words — Function.

words = cleanup ∘ to_letters ∘ to_sentences

Composite function which splits its input into sentences, then the sentences into words, and then removes special characters. Please note that dots and commas are not removed.

source

Tokenizer.to_words — Method.

to_words(tokens::Vector{String}; keeppunctuation=true)

Split all of the tokens in tokens into individual words by whitespace. If keeppunctuation is true, all of the special characters are preserved (and thus "glued" to the preceding/following word).

source